KFSPTS-14681 Fix transaction handling in RASS job
