+++ src/main/resources/edu/cornell/kfs/rass/cu-spring-rass.xml
@@ -64,20 +64,20 @@
 			p:batchInputFileService-ref="batchInputFileService"
 			p:batchInputFileType-ref="rassXmlInputFileType"
 			p:fileStorageService-ref="fileStorageService"
-			p:rassRoutingService-ref="rassRoutingService"
-			p:routeHeaderService-ref="enDocumentRouteHeaderService"
+			p:rassUpdateService-ref="rassUpdateService"
 			p:agencyDefinition-ref="agencyDefinition"
 			p:proposalDefinition-ref="proposalDefinition"
-			p:awardDefinition-ref="awardDefinition"
-			p:documentStatusCheckDelayMillis="1000"
-			p:maxStatusCheckAttempts="60"/>
+			p:awardDefinition-ref="awardDefinition"/>
 
-	<bean id="rassRoutingService" parent="rassRoutingService-parentBean"/>
-	<bean id="rassRoutingService-parentBean" abstract="true" class="edu.cornell.kfs.rass.batch.service.impl.RassRoutingServiceImpl"
+	<bean id="rassUpdateService" parent="rassUpdateService-parentBean"/>
+	<bean id="rassUpdateService-parentBean" abstract="true" class="edu.cornell.kfs.rass.batch.service.impl.RassUpdateServiceImpl"
 			p:maintenanceDocumentService-ref="maintenanceDocumentService"
 			p:documentService-ref="documentService"
 			p:dataDictionaryService-ref="dataDictionaryService"
-			p:configurationService-ref="configurationService"/>
+			p:configurationService-ref="configurationService"
+			p:routeHeaderService-ref="enDocumentRouteHeaderService"
+			p:documentStatusCheckDelayMillis="1000"
+			p:maxStatusCheckAttempts="60"/>
 
 	<bean id="rassXmlInputFileType" parent="rassXmlInputFileType-parentBean"/>
 	<bean id="rassXmlInputFileType-parentBean" abstract="true" class="edu.cornell.kfs.sys.batch.JAXBXmlBatchInputFileTypeBase"
