+++ src/test/java/edu/cornell/kfs/rass/RassTestConstants.java
@@ -23,7 +23,7 @@ public final class RassTestConstants {
     public static final String DEFAULT_PROJECT_DIRECTOR_PRINCIPAL_ID = "mo14";
 
     public static final String RASS_SERVICE_BEAN_NAME = "rassService";
-    public static final String RASS_ROUTING_SERVICE_BEAN_NAME = "rassRoutingService";
+    public static final String RASS_UPDATE_SERVICE_BEAN_NAME = "rassUpdateService";
     public static final String AGENCY_SERVICE_BEAN_NAME = "agencyService";
     public static final String AWARD_SERVICE_BEAN_NAME = "awardService";
     public static final String BUSINESS_OBJECT_SERVICE_BEAN_NAME = "businessObjectService";
